requirements.txt
numpy >= 1.9.2
matplotlib >= 1.3.1
pillow

